@@ -19,8 +19,8 @@
   "repository": "https://github.com/fr0stbyter/jspatcher",
   "devDependencies": {
     "@csound/browser": "^6.17.1",
-    "@ffmpeg/core": "^0.10.0",
-    "@ffmpeg/ffmpeg": "^0.10.1",
+    "@ffmpeg/core": "^0.11.0",
+    "@ffmpeg/ffmpeg": "^0.11.6",
     "@grame/libmusicxml": "^3.19.1",
     "@jspatcher/package-analysers": "^1.0.3",
     "@jspatcher/package-cac": "latest",
@@ -32,9 +32,8 @@
     "@jspatcher/package-std": "latest",
     "@jspatcher/package-ui": "latest",
     "@jspatcher/package-webaudio": "latest",
-    "@monaco-editor/react": "^4.2.2",
-    "@shren/faust-ui": "^1.0.4",
-    "@shren/faustwasm": "^0.0.22",
+    "@shren/faust-ui": "^1.1.2",
+    "@shren/faustwasm": "^0.0.23",
     "@shren/fftw-js": "^0.1.6",
     "@shren/guidolib": "^1.7.3",
     "@shren/worklet-loader": "^2.0.0",
@@ -42,14 +41,14 @@
     "@types/express": "^4.17.13",
     "@types/jszip": "^3.4.1",
     "@types/lz-string": "^1.3.34",
-    "@types/react": "^17.0.19",
+    "@types/react": "^18.0.27",
     "@types/react-color": "^3.0.5",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^18.0.10",
     "@types/w3c-web-serial": "^1.0.2",
     "@types/webmidi": "^2.0.6",
-    "@typescript-eslint/eslint-plugin": "^4.29.3",
-    "@typescript-eslint/parser": "^4.29.3",
-    "@typescript/vfs": "^1.3.5",
+    "@typescript-eslint/eslint-plugin": "^5.48.2",
+    "@typescript-eslint/parser": "^5.48.2",
+    "@typescript/vfs": "^1.4.0",
     "@webaudiomodules/api": "^2.0.0-alpha.4",
     "@webaudiomodules/sdk": "^0.0.10",
     "ali-oss": "^6.17.1",
@@ -58,38 +57,38 @@
     "buffer": "^6.0.3",
     "clean-webpack-plugin": "^4.0.0-alpha.0",
     "color-js": "^1.0.5",
-    "copy-webpack-plugin": "^9.0.1",
+    "copy-webpack-plugin": "^11.0.0",
     "crypto-browserify": "^3.12.0",
     "css-loader": "^6.2.0",
     "esbuild-loader": "^2.15.1",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-react": "^7.25.0",
+    "eslint": "^8.32.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-react": "^7.32.1",
     "express": "^4.17.1",
     "jszip": "^3.7.1",
     "lato-font": "^3.0.0",
-    "monaco-editor": "^0.26.1",
-    "monaco-editor-webpack-plugin": "^4.1.2",
+    "monaco-editor": "^0.34.1",
+    "monaco-editor-webpack-plugin": "^7.0.1",
     "path": "^0.12.7",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-color": "^2.19.3",
-    "react-dom": "^17.0.2",
-    "react-monaco-editor": "^0.44.0",
-    "sass": "^1.38.2",
-    "sass-loader": "^12.1.0",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.0.4",
-    "source-map-loader": "^3.0.0",
+    "react-dom": "^18.2.0",
+    "react-monaco-editor": "^0.51.0",
+    "sass": "^1.57.1",
+    "sass-loader": "^13.2.0",
+    "semantic-ui-css": "^2.5.0",
+    "semantic-ui-react": "^2.1.4",
+    "source-map-loader": "^4.0.1",
     "stream-browserify": "^3.0.0",
     "style-loader": "^3.2.1",
-    "stylelint": "^14.5.1",
-    "stylelint-config-recommended": "^7.0.0",
-    "ts-morph": "^15.1.0",
+    "stylelint": "^14.16.1",
+    "stylelint-config-recommended": "^9.0.0",
+    "ts-morph": "^17.0.1",
     "typescript": "^4.9.4",
     "util": "^0.12.4",
-    "webpack": "^5.51.1",
-    "webpack-cli": "^4.8.0",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.1",
     "window-function": "^2.1.0",
     "workbox-webpack-plugin": "^6.2.4",
     "worker-loader": "^3.0.8"
